+++ src/components/WhitePaperPage/whitePaperPage.module.scss
@@ -1,0 +1,5 @@
+.hero {
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center;
+}
