--- src/components/WhitePapers/index.js
@@ -1,3 +1,0 @@
-export * from './WhitePaperJumbotron'
-export * from './WhitePaperPage'
-export * from './WhitePaperRequestDemoForm'
